using binary protocol in the broker ingress for it to be used as an entry point for the source in the conformance tests, the changes to the broker to support this are going to be made in another commit + created source perf test results and published images, the performance gains while using the binary protocol binding vs json manipulation seems to be significant memorywise
